feat: refactor payment services by consolidating payment handling and adding cancel functionality
