+++ src/services/index.ts
@@ -2,7 +2,7 @@ export * from './books.service'
 export * from './bookSearchOptions.service'
 export * from './authors.service'
 export * from './users.service'
-export * from './payments.service'
+export * from './payments'
 export * from './orders.service'
 export * from './webhooks.service'
 export * from './news.service'
